Add files via upload
This is a rough draft of a film recommendation engine I made to help get a bit of experience working with large data sets. The data sets that the function calls is the 1,000 user data set from the MovieLens website. I intend to restructure this function in the future so that it functions as a class with methods to help compartmentalize the function's process instead of making it one large (and unwieldy) function.